refactoration+persistence
