+++ rware/persistence.h
@@ -1,0 +1,8 @@
+#pragma once
+#include <windows.h>
+#include "strsafe.h"
+#include <ShlObj.h>
+#include <iostream>
+
+int checkForRegkey();
+int setRegkey();
