--- rware/ransom_wallpaper.cpp
@@ -1,3 +1,0 @@
-#include <Windows.h>
-
-//make something that pops up a basic cmd window and a fake "ransom" text
